Minor changes to installers
Minor changes to the IIS installation so that the necessary components are installed, and not all of IIS as this cause issues when trying to create site bindings. The silent install of the Secret Server executable was also changed to use subprocess instead of os. Additional functionality was added to check the PID of the install in an effort to avoid premature deleting of the executable file.